requirements.txt
streamlit>=1.28.0
langchain>=0.3.0
langgraph>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
huggingface-hub>=0.30.0
openai>=1.50.0
groq>=0.4.0
anthropic>=0.25.0
pydantic>=2.0.0
typing-extensions>=4.7.0
apify-client>=1.11.0 